Complete response parsing for Route service
Modified the osrmc_route in the library to accept a python dictionary object
which eases the usage of the route service from the python binding.

Also added a C++ class to parse the json return value to python dictionary so
that it is easy use in the python binding.

Tested against py2 as well py3